Unit tests now work for fake frameworks.
Instead of using the hacky .fake_fw_testing copy of the framework code, it now uses the hacky UFW_FAKE_FRAMEWORK_NAME to determine the name of the framework.

Initially, it will be set to the name you choose when creating the framework target. You can change it in your build settings if you change the framework name later.