Universal builds are now done using "Archive" rather than "Build".
This speeds up dependency based builds tremendously since a universal build is not necessary in such situations.
As well, once an archive build completes, it will open the folder containing the built universal framework.
With this change, the "clean" script is no longer needed, and has been deleted.

Also fixed an issue where the script would not recognize .cpp and .mm files as compilable.