Adds support for compiled storyboards in embeddedframework
Tested with the "Fake Framework" approach.
Xcode 7.3.1 rejects an app install when an embeddedframework built with this tool (prior to my fix) is included in the project. The error is `App installation failed: Could not write to the device.`
The device logs complain about the symlink to the .storyboardc file being invalid as it attempts to install. The error is along the lines of `[MIFileManager realPathForURL:ifChildOfURL:]: Rejecting [embeddedframework/Resources/*.storyboardc] -> ../[.framework]/Resources/[*.storyboardc], as it is points outside or to the base`

Since the Resources inside the .framework file are not used, I chose to just move the .storyboardc out of the .framework and into the .embeddedframework's Resources directory (it used to symlink rather than be moved). This allows the app to install and run properly.

There may be another way to fix the symlink to iOS's liking that I don't know of - this was the first thing that worked for me.